refactor(options): keep connection test files within budgets
